+++ example/lib/pages/home/home_page/home_page.dart
@@ -70,11 +70,11 @@ class _Body extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(16),
+    return const SingleChildScrollView(
+      padding: EdgeInsets.all(16),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
-        children: const [
+        children: [
           HomeColorPalettesSection(),
           SizedBox(height: 32),
           HomeButtonsSection(),