+++ lib/src/widgets/grid/index.dart
@@ -1,2 +1,1 @@
-export 'grid_builder.dart';
 export 'static_grid.dart';