+++ lib/src/widgets/grid/static_grid.dart
@@ -1,7 +1,12 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 
-class StaticGrid extends StatelessWidget {
-  const StaticGrid({
+import 'grid_builder_util.dart';
+
+// ignore: must_be_immutable
+class StaticGrid extends StatefulWidget {
+  StaticGrid({
     Key? key,
     this.columnCount = 2,
     this.gap,
@@ -17,61 +22,155 @@ class StaticGrid extends StatelessWidget {
   final int columnCount;
   final double? gap;
   final EdgeInsets? padding;
-  final MainAxisAlignment columnMainAxisAlignment;
-  final CrossAxisAlignment columnCrossAxisAlignment;
   final MainAxisAlignment rowMainAxisAlignment;
   final CrossAxisAlignment rowCrossAxisAlignment;
-  final bool expanded;
-  final List<Widget> children;
+
+  StaticGrid.builder({
+    Key? key,
+    this.columnCount = 2,
+    this.gap,
+    this.padding,
+    this.rowMainAxisAlignment = MainAxisAlignment.start,
+    this.rowCrossAxisAlignment = CrossAxisAlignment.center,
+    required Widget Function(BuildContext, int)? itemBuilder,
+    required int? itemCount,
+    FutureOr<void> Function()? onLoadData,
+  })  : _onLoadData = onLoadData,
+        _itemCount = itemCount,
+        _itemBuilder = itemBuilder,
+        _isBuilder = true,
+        super(key: key);
+
+  StaticGrid.sliver({
+    Key? key,
+    this.columnCount = 2,
+    this.gap,
+    this.padding,
+    this.rowMainAxisAlignment = MainAxisAlignment.start,
+    this.rowCrossAxisAlignment = CrossAxisAlignment.center,
+    required Widget Function(BuildContext, int)? itemBuilder,
+    required int? itemCount,
+    int? Function(Key)? findChildIndexCallback,
+    bool? addAutomaticKeepAlives = true,
+    bool? addRepaintBoundaries = true,
+    FutureOr<void> Function()? onLoadData,
+  })  : _onLoadData = onLoadData,
+        _addRepaintBoundaries = addRepaintBoundaries,
+        _addAutomaticKeepAlives = addAutomaticKeepAlives,
+        _findChildIndexCallback = findChildIndexCallback,
+        _itemCount = itemCount,
+        _itemBuilder = itemBuilder,
+        _isSliver = true,
+        super(key: key);
+
+  // Property for StaticGrid
+  List<Widget>? children;
+  bool? expanded;
+  MainAxisAlignment? columnMainAxisAlignment;
+  CrossAxisAlignment? columnCrossAxisAlignment;
+
+  // Property for StaticGrid.builder and StaticGrid.sliver
+  Widget Function(BuildContext ontext, int index)? _itemBuilder;
+  int? _itemCount;
+
+  // Property for StaticGrid.sliver
+  int? Function(Key)? _findChildIndexCallback;
+  bool? _addAutomaticKeepAlives;
+  bool? _addRepaintBoundaries;
+
+  bool _isBuilder = false;
+  bool _isSliver = false;
+  FutureOr<void> Function()? _onLoadData;
+
+  @override
+  State<StaticGrid> createState() => _StaticGridState();
+}
+
+class _StaticGridState extends State<StaticGrid> {
+  bool _isLoading = false;
+
+  Future<void> _onLoadData() async {
+    if (_isLoading) {
+      return;
+    }
+
+    _isLoading = true;
+    await widget._onLoadData?.call();
+    _isLoading = false;
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      padding: padding,
-      child: Column(
-        crossAxisAlignment: columnCrossAxisAlignment,
-        mainAxisAlignment: columnMainAxisAlignment,
-        children: _createRows(),
-      ),
-    );
-  }
-
-  List<Widget> _createRows() {
-    final List<Widget> rows = [];
-    final rowCount = (children.length / columnCount).ceil();
-
-    for (int rowIndex = 0; rowIndex < rowCount; rowIndex++) {
-      final row = Row(
-        crossAxisAlignment: rowCrossAxisAlignment,
-        mainAxisAlignment: rowMainAxisAlignment,
-        children: _createCells(rowIndex),
+    if (!widget._isBuilder && !widget._isSliver) {
+      final rowCount = (widget.children!.length / widget.columnCount).ceil();
+      return Container(
+        padding: widget.padding,
+        child: Column(
+          crossAxisAlignment: widget.columnCrossAxisAlignment!,
+          mainAxisAlignment: widget.columnMainAxisAlignment!,
+          children: List.generate(
+            rowCount,
+            (index) => GridBuilderUtil.createRow(
+              gap: widget.gap,
+              rowIndex: index,
+              itemCount: widget.children!.length,
+              rowCount: rowCount,
+              columnCount: widget.columnCount,
+              rowMainAxisAlignment: widget.rowMainAxisAlignment,
+              rowCrossAxisAlignment: widget.rowCrossAxisAlignment,
+              builder: (context, index) => widget.children![index],
+            ),
+          ),
+        ),
       );
-
-      rows.add(expanded ? Expanded(child: row) : row);
-
-      if (rowIndex != rowCount - 1) {
-        rows.add(SizedBox(height: gap));
-      }
     }
 
-    return rows;
-  }
+    if (widget._isBuilder) {
+      final rowCount = (widget._itemCount! / widget.columnCount).ceil() +
+          (widget._onLoadData != null ? 1 : 0);
 
-  List<Widget> _createCells(int rowIndex) {
-    final List<Widget> columns = [];
-    for (int columnIndex = 0; columnIndex < columnCount; columnIndex++) {
-      final cellIndex = rowIndex * columnCount + columnIndex;
-      if (cellIndex <= children.length - 1) {
-        columns.add(Expanded(child: children[cellIndex]));
-      } else {
-        columns.add(Expanded(child: Container()));
-      }
-
-      if (columnIndex != columnCount - 1) {
-        columns.add(SizedBox(width: gap));
-      }
+      return ListView.builder(
+        padding: widget.padding,
+        itemBuilder: (context, index) => GridBuilderUtil.createRow(
+          gap: widget.gap,
+          rowIndex: index,
+          itemCount: widget._itemCount!,
+          rowCount: rowCount,
+          columnCount: widget.columnCount,
+          rowMainAxisAlignment: widget.rowMainAxisAlignment,
+          rowCrossAxisAlignment: widget.rowCrossAxisAlignment,
+          builder: widget._itemBuilder!,
+          onLoadData: widget._onLoadData != null ? _onLoadData : null,
+        ),
+        itemCount: rowCount,
+      );
     }
 
-    return columns;
+    if (widget._isSliver) {
+      final rowCount = (widget._itemCount! / widget.columnCount).ceil() +
+          (widget._onLoadData != null ? 1 : 0);
+
+      return SliverList(
+        delegate: SliverChildBuilderDelegate(
+          (context, index) => GridBuilderUtil.createRow(
+            gap: widget.gap,
+            rowIndex: index,
+            itemCount: widget._itemCount!,
+            rowCount: rowCount,
+            columnCount: widget.columnCount,
+            rowMainAxisAlignment: widget.rowMainAxisAlignment,
+            rowCrossAxisAlignment: widget.rowCrossAxisAlignment,
+            builder: widget._itemBuilder!,
+            onLoadData: widget._onLoadData != null ? _onLoadData : null,
+          ),
+          childCount: rowCount,
+          findChildIndexCallback: widget._findChildIndexCallback,
+          addAutomaticKeepAlives: widget._addAutomaticKeepAlives!,
+          addRepaintBoundaries: widget._addRepaintBoundaries!,
+        ),
+      );
+    }
+
+    return const SizedBox();
   }
 }