+++ lib/src/widgets/infinite_scroll/index.dart
@@ -1,2 +1,1 @@
 export 'infinite_scroll.dart';
-export 'sliver_infinite_scroll.dart';